working on fixing bogus mbr partition tables...

+++ newcase.h
@@ -20,8 +20,8 @@
 */
 
 void ProcessVolume(QString evidstring);
-int ParseVolume(QString estring, qint64 imgsize, QList<int> *offlist, QList<int> *sizlist);
-void ParseExtendedPartition(QString estring, uint32_t primaryextendedoffset, uint32_t offset, uint32_t size, QList<int>* offlist, QList<int>* sizelist);
+int ParseVolume(QString estring, qint64 imgsize, QList<qint64> *offlist, QList<qint64> *sizlist);
+void ParseExtendedPartition(QString estring, uint32_t primaryextendedoffset, uint32_t offset, uint32_t size, QList<qint64>* offlist, QList<qint64>* sizelist);
 int GetFileSystemType(QString estring, off64_t offset);
 /*
 //int GetFileSystemType(QString estring);
